Replaced TokenIterator class with Lexer class.
This replacement aims to acknowledge and solve every issue we had with
the previous design. Keywords and Terminals are now defined in the same
place via macros. We no longer do a 2-pass when tokenizing and parsing.
The current token is now saved in the lexer rather than in the parser.
There is now only one Token class rather than three.