Completed the dominance algorithm.
Instruction, BasicBlock, and IntermediateRepresentation classes are
completed. Added the ability to create basic blocks from the IR. When
this occurs, they're automatically added in reverse postorder and their
immediate dominator is calculated. Added a function that prints a dot
graph language representation of the IR.

BasicBlock and Instruction objects are completely handled by the IR. The
Parser should only need an integer representing the index where the
BasicBlock its working with exists in the IR's basic_block vector.