MAJOR refactor of the codebase.
- Moved token related code out of lexer.cpp and into token.cpp.
- Added additional types to the Token class' variant member. Debugging
is now far easier since it properly displays keyword identifiers and
user identifiers rather than random integers. Token's are now of type
int (for CONSTANTs), Terminal, Keyword, and ident_t (for IDENTIFIERs).
- Fixed concerning issue with how the Lexer handled Keywords. The
KEYWORD enum now starts at 1 (due to the addition of DUMMY).
- Made an opcode.hpp file and moved all opcode related code there.
- For consistency, moved private fields of all classes to the bottom.
- Added helper functions to the IntermediateRepresentation class to
remove copy-pasted code which is prone to error.
- Added helper functions to the Parser class to remove copy-pasted code
which is prone to error.
- Heavily refactored the Parser class, adding additional helper
functions along with splitting up big functions into a main function and
several helpers specifically for it.
- Parser class should now handle far more Tiny syntax errors.